Refactor routes and startup
